app/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
python-binance
pandas
plotly
python-coinmarketcap